Update build system
We dropped prepack as it's not currently maintained and it's not really
useful for that project.
Instead of relying completely on Rollup with plugins, we now use
TypeScript directly to transpile code, and then Rollup to bundle the
library as a UMD package. This workflow permits us to have good
sourcemaps and more robust builds.
We have also dropped Just to handle our tasks, as NPM scripts are enough
and straightforward for what we need.